+++ converters/manpage/src/inlines.rs
@@ -597,10 +597,13 @@ impl<'a, W: Write> ManpageVisitor<'a, '_, W> {
                 write!(self.writer_mut(), "{closing_quote}")?;
                 Ok(())
             }
-            XrefDisplay::FullEmphasized(prefix, inlines, _scope) => {
-                let prefix = manify(&prefix, EscapeMode::Normalize);
-                let separator = manify(", ", EscapeMode::Normalize);
-                write!(self.writer_mut(), "{prefix}{separator}\\fI")?;
+            XrefDisplay::Emphasized(prefix, inlines, _scope) => {
+                if let Some(prefix) = prefix {
+                    let prefix = manify(&prefix, EscapeMode::Normalize);
+                    let separator = manify(", ", EscapeMode::Normalize);
+                    write!(self.writer_mut(), "{prefix}{separator}")?;
+                }
+                write!(self.writer_mut(), "\\fI")?;
                 self.visit_inline_nodes(traversal, inlines)?;
                 write!(self.writer_mut(), "\\fP")?;
                 Ok(())